🚚 Rename: 그룹룰렛리스트 폴더 생성 및 하위로 이동
