+++ src/components/mainPage/groupRouletteList/groupRouletteList.module.scss
